Refactor to fsd 2.1, move logic to pages
