+++ src/app/prodivers/router/RouterProvider.tsx
@@ -1,0 +1,6 @@
+import { RouterProvider as ReactRouterProvider } from 'react-router-dom'
+import { appRouter } from './appRouter'
+
+export function RouterProvider() {
+  return <ReactRouterProvider router={appRouter()} />
+}
