+++ src/pages/main/index.ts
@@ -1,1 +1,2 @@
 export { MainPage } from './ui/Page/Page'
+export { productPopularListHandlers } from './api/__mocks__/productPopularListHandlers'
