+++ src/pages/product/index.ts
@@ -1,1 +1,2 @@
 export { ProductPage } from './ui/Page/Page'
+export { productDetailsHandlers } from './api/__mocks__/productDetailsHandlers'
