+++ src/pages/product/lib/mapProductDetails.ts
