+++ src/pages/product/lib/transformProductDetailsToProduct.ts
