+++ src/pages/product/model/types.ts
