+++ src/pages/product/ui/ProductDetails/ProductDetails.module.css
