--- src/shared/ui/Layout/Layout.tsx
@@ -1,41 +1,0 @@
-import type { ReactNode } from 'react'
-import { Outlet, ScrollRestoration } from 'react-router-dom'
-import css from './Layout.module.css'
-
-type Props = {
-  navbarSlot?: ReactNode
-  headerSlot: ReactNode
-  bottomSlot?: ReactNode
-  announcementSlot?: ReactNode
-  sidebarSlot?: ReactNode
-}
-
-export function Layout(props: Props) {
-  return (
-    <div className={css.root}>
-      {props.announcementSlot}
-      {props.navbarSlot}
-      {props.headerSlot}
-      <div className={css.container}>
-        <div className={css.content}>
-          <Outlet />
-        </div>
-        {props.sidebarSlot && (
-          <aside className={css.sidebar}>{props.sidebarSlot}</aside>
-        )}
-      </div>
-      <footer className={css.footer}>
-        <div className="text_sm">
-          {new Date().getFullYear()}
-          , see source code on
-          {' '}
-          <a href="https://github.com/noveogroup-amorgunov/nukeapp">
-            github/noveogroup-amorgunov/nukeapp
-          </a>
-        </div>
-      </footer>
-      {props.bottomSlot}
-      <ScrollRestoration />
-    </div>
-  )
-}
