+++ src/shared/ui/index.ts
@@ -1,5 +1,4 @@
 export { Announcement } from './Announcement/Announcement'
-export { Layout } from './Layout/Layout'
 export { Icon } from './Icon/Icon'
 export { Button } from './Button/Button'
 export { Input } from './Input/Input'
