--- src/widgets/CategoryPopularList/index.ts
@@ -1,1 +1,0 @@
-export { CategoryPopularList } from './ui/CategoryPopularList/CategoryPopularList'
