+++ src/widgets/Layout/ui/Layout/Layout.module.css
@@ -1,0 +1,49 @@
+@import url('/src/shared/breakpoints.css');
+
+.root {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  padding: 64px 24px 16px;
+  min-height: 100vh;
+
+  @media (--s-viewport) {
+    padding: 64px 8px 16px;
+  }
+
+  @media (--xs-viewport) {
+    padding-top: 84px;
+  }
+}
+
+.container {
+  display: flex;
+  flex-direction: row;
+  gap: 12px;
+  flex-grow: 1;
+
+  @media (--s-viewport) {
+    flex-direction: column-reverse;
+  }
+}
+
+.content {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  flex-grow: 1;
+}
+
+.sidebar {
+  min-width: 300px;
+  display: flex;
+
+  @media (--s-viewport) {
+    min-width: 100%;
+  }
+}
+
+.footer {
+  text-align: center;
+  padding-top: 40px;
+}
