--- src/widgets/LayoutHeader/index.ts
@@ -1,1 +1,0 @@
-export { LayoutHeader } from './ui/LayoutHeader/LayoutHeader'
