--- src/widgets/ProductDetails/index.ts
@@ -1,2 +1,0 @@
-export { ProductDetails } from './ui/ProductDetails'
-export { useProductDetailsQuery } from './api/widgetProductDetailsApi'
