--- src/widgets/ProductPopularList/api/__mocks__/widgetProductPopularListHandlers.ts
@@ -1,17 +1,0 @@
-import { HttpResponse, delay, http } from 'msw'
-import { env } from '@/shared/lib'
-import { __serverDatabase } from '@/shared/lib/server'
-
-export const widgetProductPopularListHandlers = [
-  http.get(`${env.VITE_API_ENDPOINT}/products/popular`, async () => {
-    const products = __serverDatabase.product.findMany({
-      where: { popular: { equals: true } },
-    })
-
-    await delay(env.VITE_API_DELAY)
-    return HttpResponse.json(
-      products,
-      { status: 200 },
-    )
-  }),
-]
