--- src/widgets/ProductPopularList/index.ts
@@ -1,1 +1,0 @@
-export { ProductPopularList } from './ui/ProductPopularList'
